fix: drop `fontWeight` besides `bold`
Following discussion today, it was decided that we only need to define
a single weight (besides `normal`), `bold`, which would use the `600`
weight that is recommended by the brand team (instead of the default
`700`).

This allows developers to ignore the fact that we technically never want
to use "bold", and instead always want "semibold".

We also style the semantic `b` and `strong` elements with the correct
weight.