refactor: disable responsive variants globally
Rather than configuring each plugin's variants to remove the responsive
one, we can just configure the whole theme to have no "screens" config,
which will ensure that no responsive CSS helpers are generated.

This also has the side-effect of enabling the default variants for all
enabled plugins, which is what we want.

Closes #8