fix: use `rem` units for line height
The description in our Typography guidelines mentions pixel values for
each corresponding `line-height` value, since that is typically how
design tools work.

If you configure `line-height` without a unit, the amount of space is
relative to the font size of the text being styled, _not_ the document
as a whole. That means that this

    <p class="text-base leading-base">Whatever</p>

And this

    <p class="text-lg leading-base">Whatever</p>

Actually have different `line-height` values!

Instead, we should base the line height on the font size of the document
itself (by using `rem`) so that values are more predictable and match
a value easier for designers to work with.