feat: generate stable Lorem Ipsum examples
This replaces the default `Math.random` implementation within
`LoremIpsum` with an alternate implementation that can be given a seed,
so that it always generates the same text examples.

This should stabilize our Percy snapshots, so that we end up with the
same text generated in each of them. Without this behavior, each
snapshot would get invalidated every time, just from the dummy text
changing.