feat: enable max/min height/width plugins
The max- and min- height and width plugins from Tailwind actually are
not configured based on a scale at all. Because of this, it's really
unlikely that we would have an opinion on what sizes are made available
to use with them. They should be safe for us to enable even before
we have the actual height/width size scale defined.