fix: improve states when used with disabled appearance
The `<button>` tag automatically makes certain combinations of states impossible, so explicit CSS was not written to account for them, specifically around the button being disabled and thus not being able to be active or focused.

However, since the `<a>` does _not_ provide a `disabled` attribute that we can use, we need to ensure that it does not also take on the active or focused styles when the disabled appearance is present explicitly.