fix: allow an `sr-only` element within an icon-only button
This allows us to use the `sr-only` utility class from Tailwind to add text
within an icon-only button for accessibility purposes.

This is useful because `aria-label` has limitations that rendering directly
into the DOM does not.

The previous CSS would add the margin next to the icon even if the text was
marked as `sr-only`. Now that class is specifically detected and the spacing
added only if the text does _not_ have the `sr-only` class.

Closes #257